pyproject.toml
[project]
name = "dotfiles"
version = "8.0.23"
description = "Personal development laboratory"
authors = [{ name = "Itzalak", email = "itzalak@tuta.io" }]
readme = "README.md"
keywords = ["dotfiles", "archlinux", "hyprland", "macos"]
requires-python = ">=3.13"
dependencies = [
  "pytest >= 8.3.3",
  "pyyaml >= 6.0.2",
  "python-dotenv >= 1.0.1",
  "toml >= 0.10.2",
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.commitizen]
name = "cz_customize"
tag_format = "v$major.$minor.$patch"
version_scheme = "semver"
version = "8.0.10"
update_changelog_on_bump = true
version_provider = "uv"

[tool.uv]
required-version = ">=0.7.3"

[tool.commitizen.customize]
schema = "<type>: <body>"
schema_pattern = "(bump|break|feat|chore|tools|docs|fix):(\\s.*)"
bump_pattern = "^(bump|break|feat|tools|fix|chore)"
bump_map = { "break" = "MAJOR", "feat" = "MINOR", "fix" = "PATCH", "tools" = "PATCH", "chore" = "PATCH", "docs" = "PATCH" }
change_type_order = ["break", "feat", "fix", "tools", "chore", "docs"]
commit_parser = "^(?P<change_type>break|feat|chore|tools|docs|fix):\\s(?P<message>.*)?"
changelog_pattern = "^(break|feat|chore|tools|docs|fix)?(!)?"
change_type_map = { "break" = "Breaking Change", "feat" = "Changed", "fix" = "Fixed", "tools" = "Tools", "chore" = "Changed", "docs" = "Notes" }
